Renamed background service.
